add ToGenericError instances for the type checker errors
