+++ src/MiniJuvix/Syntax/MicroJuvix/Error/Pretty.hs
@@ -1,10 +1,8 @@
 module MiniJuvix.Syntax.MicroJuvix.Error.Pretty
-  ( module MiniJuvix.Syntax.MicroJuvix.Error.Pretty.Base,
+  ( module MiniJuvix.Syntax.MicroJuvix.Error.Types,
     module MiniJuvix.Syntax.MicroJuvix.Error.Pretty.Ansi,
-    module MiniJuvix.Syntax.MicroJuvix.Error.Pretty.Text,
   )
 where
 
 import MiniJuvix.Syntax.MicroJuvix.Error.Pretty.Ansi
-import MiniJuvix.Syntax.MicroJuvix.Error.Pretty.Base
-import MiniJuvix.Syntax.MicroJuvix.Error.Pretty.Text
+import MiniJuvix.Syntax.MicroJuvix.Error.Types
