+++ src/MiniJuvix/Syntax/MicroJuvix/Error/Pretty/Ann.hs
@@ -1,0 +1,7 @@
+module MiniJuvix.Syntax.MicroJuvix.Error.Pretty.Ann where
+
+import MiniJuvix.Syntax.MicroJuvix.Pretty.Ann qualified as Micro
+
+data Eann
+  = Highlight
+  | MicroAnn Micro.Ann
