--- src/MiniJuvix/Syntax/MicroJuvix/Error/Pretty/Text.hs
@@ -1,9 +1,0 @@
-module MiniJuvix.Syntax.MicroJuvix.Error.Pretty.Text where
-
-import MiniJuvix.Prelude
-import MiniJuvix.Syntax.MicroJuvix.Error.Pretty.Base
-import Prettyprinter
-import Prettyprinter.Render.Text
-
-renderText :: SimpleDocStream Eann -> Text
-renderText = renderStrict
